# docs/api/events-api.openapi.yml
openapi: 3.0.3
info:
  title: BATbern Events API
  version: 1.0.0
  description: |
    Consolidated Events API for BATbern Event Management Platform.

    This API consolidates 130 fragmented endpoints into 25 efficient endpoints using:
    - **Resource Expansion**: `?include=` parameter for sub-resources
    - **Rich Filtering**: JSON filter syntax for complex queries
    - **Pagination**: Standard page/limit parameters
    - **Caching**: Caffeine in-memory cache (15min TTL)

    **Story**: 1.15a.1 - Events API Consolidation

    **Performance Targets**:
    - Event list (no includes): <100ms (P95)
    - Event detail (basic): <150ms (P95)
    - Event detail (all includes): <500ms (P95)
    - Cached response: <50ms (P95)

  contact:
    name: BATbern Platform Team
    email: platform@batbern.ch
  license:
    name: Proprietary
servers:
  - url: https://api.batbern.ch/api/v1
    description: Production
  - url: https://api.staging.batbern.ch/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Development
...
paths:
...
  /events/{eventCode}:
    get:
      tags:
        - Events
      summary: Get event details
      description: |
        Retrieve detailed information about a specific event with optional resource expansion.

        **Acceptance Criteria**: AC2
        **Story**: 1.16.2 - Uses eventCode (meaningful ID) instead of UUID

        **Include Options**:
        - `venue`: Venue details with address and capacity
        - `sessions`: Event sessions with schedule (speakers included in sessions)
        - `topics`: Event topics and categories
        - `workflow`: Workflow state and history
        - `registrations`: Registration counts and status
        - `catering`: Catering arrangements
        - `team`: Team assignments
        - `publishing`: Publishing details
        - `notifications`: Notification settings
        - `analytics`: Event analytics and metrics

        **Performance**:
        - Basic (no includes): <150ms (P95)
        - All includes: <500ms (P95)
        - Cached: <50ms (P95)

        **Caching**: 15-minute TTL with automatic invalidation on updates

      operationId: getEvent
      parameters:
        - name: eventCode
          in: path
          description: Event code in format BATbern{number}
          required: true
          schema:
            type: string
            pattern: '^BATbern[0-9]+$'
            example: BATbern142
        - name: include
          in: query
          description: Comma-separated list of resources to include
          required: false
          schema:
            type: string
            example: 'venue,sessions,workflow'
      responses:
        '200':
          description: Event details retrieved successfully
          headers:
            X-Cache-Status:
              description: Cache hit status (HIT or MISS)
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDetail'
              examples:
                basicEvent:
                  summary: Basic event (no includes)
                  value:
                    eventCode: BATbern142
                    title: Spring Conference 2025
                    eventNumber: 142
                    date: 2025-05-15T09:00:00Z
                    registrationDeadline: 2025-05-01T23:59:59Z
                    venueName: Kursaal Bern
                    venueAddress: Kornhausstrasse 3, 3013 Bern
                    venueCapacity: 500
                    workflowState: PUBLISHED
                    organizerUsername: john.doe
                    currentAttendeeCount: 187
                    description: Cloud Native Architecture
                expandedEvent:
                  summary: Event with venue
                  value:
                    eventCode: BATbern142
                    title: Spring Conference 2025
                    eventNumber: 142
                    date: 2025-05-15T09:00:00Z
                    registrationDeadline: 2025-05-01T23:59:59Z
                    venueName: Kursaal Bern
                    venueAddress: Kornhausstrasse 3, 3013 Bern
                    venueCapacity: 500
                    workflowState: PUBLISHED
                    organizerUsername: john.doe
                    currentAttendeeCount: 187
                    description: Cloud Native Architecture
                    venue:
                      id: ven-001
                      name: Kursaal Bern
                      address: Kornhausstrasse 3, 3013 Bern
                      capacity: 500
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Events
      summary: Update event (full replacement)
      description: |
        Replace all event data with new data.

        **Acceptance Criteria**: AC4
        **Story**: 1.16.2 - Uses eventCode (meaningful ID) instead of UUID

        **Cache Invalidation**: All event caches cleared on update

      operationId: updateEvent
      parameters:
        - name: eventCode
          in: path
          description: Event code in format BATbern{number}
          required: true
          schema:
            type: string
            pattern: '^BATbern[0-9]+$'
            example: BATbern142
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Events
      summary: Partially update event
      description: |
        Update specific fields of an event without replacing all data.

        **Acceptance Criteria**: AC5
        **Story**: 1.16.2 - Uses eventCode (meaningful ID) instead of UUID

        **Cache Invalidation**: All event caches cleared on update

      operationId: patchEvent
      parameters:
        - name: eventCode
          in: path
          description: Event code in format BATbern{number}
          required: true
          schema:
            type: string
            pattern: '^BATbern[0-9]+$'
            example: BATbern142
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchEventRequest'
            examples:
              updateTitle:
                summary: Update only title
                value:
                  title: Spring Conference 2025 - Updated
      responses:
        '200':
          description: Event patched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Events
      summary: Delete event
      description: |
        Delete an event and all associated data.

        **Acceptance Criteria**: AC6
        **Story**: 1.16.2 - Uses eventCode (meaningful ID) instead of UUID

        **Cascade Deletion**: Deletes all sessions, registrations, and related data
        **Cache Invalidation**: All event caches cleared on deletion

      operationId: deleteEvent
      parameters:
        - name: eventCode
          in: path
          description: Event code in format BATbern{number}
          required: true
          schema:
            type: string
            pattern: '^BATbern[0-9]+$'
            example: BATbern142
      responses:
        '204':
          description: Event deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /events/{eventCode}/timetable:
    get:
      tags:
        - Sessions
      summary: Get the authoritative timetable for an event
      description: |
        Returns the complete chronological slot grid for the event, including both
        structural slots (MODERATION, BREAK, LUNCH) and droppable speaker slots
        (SPEAKER_SLOT).  Structural slots are enriched with DB session slugs when
        structural sessions have been generated.  Speaker slots show which speaker
        session (if any) has been assigned to each position.

        This endpoint is the single source of truth for the slot-assignment UI —
        both the grid display and the auto-assign algorithm use these positions.

        **Access**: ORGANIZER role required.
      operationId: getEventTimetable
      security:
        - BearerAuth: []
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
          example: BATbern142
      responses:
        '200':
          description: Timetable successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimetableResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    TimetableSlot:
      type: object
      description: |
        A single slot in the event timetable. SPEAKER_SLOT entries represent droppable
        speaker positions (the implicit gaps between structural sessions). All other types
        map to persisted structural sessions (moderation, break, lunch).
      required:
        - type
        - startTime
        - endTime
      properties:
        type:
          type: string
          enum: [MODERATION, BREAK, LUNCH, SPEAKER_SLOT]
          description: Slot type — determines rendering and whether the slot is droppable
          example: SPEAKER_SLOT
        startTime:
          type: string
          format: date-time
          description: Slot start time (UTC ISO-8601)
          example: "2025-06-15T07:05:00Z"
        endTime:
          type: string
          format: date-time
          description: Slot end time (UTC ISO-8601)
          example: "2025-06-15T07:50:00Z"
        title:
          type: string
          nullable: true
          description: Human-readable label for structural slots; null for SPEAKER_SLOT
          example: "Kaffee-Pause"
        slotIndex:
          type: integer
          nullable: true
          description: >
            1-based global index for SPEAKER_SLOT entries across the full event day
            (never reset when switching AM → PM block). Null for structural slots.
          example: 1
        sessionSlug:
          type: string
          nullable: true
          description: >
            Session slug of the structural DB session occupying this slot.
            Null when no structural sessions have been generated yet, or for SPEAKER_SLOT.
          example: "moderation-start-abc"
        assignedSessionSlug:
          type: string
          nullable: true
          description: >
            Session slug of the speaker session assigned to this SPEAKER_SLOT.
            Null if the slot is free, or for structural slots.
          example: "speaker-xyz"
    TimetableResponse:
      type: object
      description: Full timetable for an event including slot grid and unassigned sessions.
      required:
        - slots
        - unassignedSessions
      properties:
        slots:
          type: array
          description: All slots in chronological order (structural and speaker slots interleaved)
          items:
            $ref: '#/components/schemas/TimetableSlot'
        unassignedSessions:
          type: array
          description: Speaker sessions (non-structural) with no startTime assigned yet
          items:
            $ref: '#/components/schemas/SessionResponse'

    # ── Watch App Schemas (W2.4) ─────────────────────────────────────────────
...
    Event:
      type: object
      required:
        - eventCode
        - title
        - eventNumber
        - date
        - registrationDeadline
        - venueName
        - venueAddress
        - venueCapacity
        - organizerUsername
      properties:
        eventCode:
          type: string
          pattern: '^BATbern[0-9]+$'
          example: BATbern142
          description: |
            Human-readable event identifier in format "BATbern{number}".
            Story 1.16.2: Public API uses meaningful IDs, not UUIDs.
        title:
          type: string
          maxLength: 255
          example: Spring Conference 2025
        eventNumber:
          type: integer
          example: 142
          description: Sequential event number (e.g., BATbern #142)
        date:
          type: string
          format: date-time
          example: 2025-05-15T09:00:00Z
          description: Event date
        registrationDeadline:
          type: string
          format: date-time
          example: 2025-05-01T23:59:59Z
        venueName:
          type: string
          maxLength: 255
          example: Kursaal Bern
        venueAddress:
          type: string
          example: Kornhausstrasse 3, 3013 Bern
        venueCapacity:
          type: integer
          example: 500
        organizerUsername:
          type: string
          pattern: '^[a-z]+\.[a-z]+(\.[0-9]+)?$'
          example: john.doe
          description: |
            Username of the event organizer in format "firstname.lastname" or "firstname.lastname.2" for collisions.
            Story 1.16.2: Public API uses meaningful IDs (usernames), not UUIDs.
        currentAttendeeCount:
          type: integer
          example: 0
          default: 0
        publishedAt:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: string
          description: JSONB metadata
          example: '{"theme":"cloud-native","category":"technology"}'
        description:
          type: string
          example: Cloud Native Architecture and Best Practices
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        themeImageUrl:
          type: string
          nullable: true
          format: uri
          maxLength: 1000
          example: "https://d123456.cloudfront.net/logos/2025/events/BATbern57/theme-abc123.png"
          description: |
            CloudFront URL for event theme image.
            Story 2.5.3a: Event Theme Image Upload
        themeImageUploadId:
          type: string
          nullable: true
          maxLength: 100
          example: "abc123-def456"
          description: |
            Upload ID from three-phase upload pattern (internal reference).
            Story 2.5.3a: Event Theme Image Upload
            This is returned so the frontend can track what was uploaded.
        topicCode:
          type: string
          nullable: true
          pattern: '^[a-z0-9-]+$'
          description: |
            Topic code (slug-format identifier) of the selected topic.
            Story 5.2: Topic Selection & Speaker Brainstorming
            ADR-003: Meaningful Identifiers - use topicCode instead of UUID
          example: 'cloud-native-security-2024'
        eventType:
          $ref: '#/components/schemas/EventType'
          description: |
            Event type (FULL_DAY, AFTERNOON, EVENING).
            Story 5.1: Event Type Definition
        typicalStartTime:
          type: string
          format: time
          description: |
            Typical start time for this event type (from event_types table).
            Combined with event date for display purposes.
          example: "09:00"
        typicalEndTime:
          type: string
          format: time
          description: |
            Typical end time for this event type (from event_types table).
            Combined with event date for display purposes.
          example: "16:00"
        workflowState:
          $ref: '#/components/schemas/EventWorkflowState'
          description: |
            Current workflow state (16-step workflow).
            Story 5.1a: Event Workflow State Machine
        # Optional resource expansion fields (Story BAT-109: Archive Browsing)
        # These fields are populated only when requested via ?include parameter
        topic:
          type: object
          nullable: true
          description: |
            Expanded topic details (only included when ?include=topics).
            Story BAT-109: Archive browsing with topic expansion
          properties:
            code:
              type: string
              example: "cloud-native-security-2024"
            name:
              type: string
              example: "Cloud Native Security"
            description:
              type: string
              example: "Security patterns for cloud-native applications"
            category:
              type: string
              enum: [technical, management, soft_skills, industry_trends, tools_platforms]
              example: "technical"
        venue:
          $ref: '#/components/schemas/Venue'
          nullable: true
          description: |
            Expanded venue details (only included when ?include=venue).
            Story BAT-109: Archive browsing with venue expansion
        sessions:
          type: array
          nullable: true
          description: |
            Expanded session list with speakers (only included when ?include=sessions or ?include=speakers).
            Story BAT-109: Archive browsing with session expansion
          items:
            $ref: '#/components/schemas/Session'
    EventDetail:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            venue:
              $ref: '#/components/schemas/Venue'
            sessions:
              type: array
              items:
                $ref: '#/components/schemas/Session'
            currentPublishedPhase:
              type: string
              nullable: true
              enum: [TOPIC, SPEAKERS, AGENDA]
              description: |
                Progressive publishing phase for watch app (Story W1.2)
                - TOPIC: Event topic selected, speakers TBA
                - SPEAKERS: Speaker list published
                - AGENDA: Full session details published
    Venue:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        capacity:
          type: integer
...
    Session:
      type: object
      description: |
        Story 1.16.2: Uses sessionSlug and eventCode as meaningful IDs

        **Placeholder Session Pattern (V21):**
        Sessions can be created as placeholders when speakers accept invitations.
        Placeholder sessions have title but null sessionType/startTime/endTime.
        Timing is assigned later during slot assignment workflow (Story 5.7).
      required:
        - sessionSlug
        - eventCode
        - title
      properties:
        sessionSlug:
          type: string
          description: URL-friendly slug generated from session title
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
          example: keynote-opening-session
          maxLength: 200
        eventCode:
          type: string
          description: Event code this session belongs to
          pattern: '^BATbern[0-9]+$'
          example: BATbern142
        title:
          type: string
          maxLength: 255
        description:
          type: string
        sessionType:
          type: string
          nullable: true
          description: Session type - null for placeholder sessions, assigned during agenda planning
          enum: [keynote, presentation, workshop, panel_discussion, networking, break, lunch, moderation]
          example: keynote
        startTime:
          type: string
          format: date-time
          nullable: true
          description: Session start time - null for placeholder sessions, assigned during slot assignment
        endTime:
          type: string
          format: date-time
          nullable: true
          description: Session end time - null for placeholder sessions, assigned during slot assignment
        room:
          type: string
          maxLength: 100
          example: Main Hall
        capacity:
          type: integer
          example: 200
        language:
          type: string
          maxLength: 10
          example: de
          default: de
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        speakers:
          type: array
          description: |
            Story 1.15a.1b: Session speakers with enriched User data
            List of speakers assigned to this session with their roles
          items:
            $ref: '#/components/schemas/SessionSpeaker'
        materialsStatus:
          type: string
          description: |
            Material submission workflow status
            - NONE: No materials submitted yet (initial state)
            - pending: Materials submitted, awaiting review
            - in_review: Materials under review
            - approved: Materials approved
            - requires_changes: Materials need changes
            - rejected: Materials rejected
            - revision_submitted: Revised materials submitted
            - COMPLETE: Materials finalized and approved (final state)
          enum: [NONE, pending, in_review, approved, requires_changes, rejected, revision_submitted, COMPLETE]
          example: pending
    SessionSpeaker:
      type: object
      description: |
        Story 1.15a.1b: Session speaker with enriched User data
        Combines SessionUser entity (role, confirmation) with User entity (name, company, photo)
      required:
        - username
        - firstName
        - lastName
        - speakerRole
        - isConfirmed
      properties:
        username:
          type: string
          description: User's username (public identifier from User entity)
          pattern: '^[a-z]+\.[a-z]+(\.[0-9]+)?$'
          example: john.doe
          maxLength: 100
        firstName:
          type: string
          description: Speaker's first name (from User entity)
          example: John
          maxLength: 100
        lastName:
          type: string
          description: Speaker's last name (from User entity)
          example: Doe
          maxLength: 100
        company:
          type: string
          description: Speaker's company name (from User.companyId)
          example: GoogleZH
          maxLength: 12
        profilePictureUrl:
          type: string
          format: uri
          description: Speaker's profile picture URL (from User entity)
          example: https://cdn.batbern.ch/logos/2025/users/john.doe/profile.jpg
        bio:
          type: string
          description: Speaker's biography (from User entity)
          maxLength: 2000
          example: Software engineer passionate about cloud architecture and security
        speakerRole:
          type: string
          enum: [PRIMARY_SPEAKER, CO_SPEAKER, MODERATOR, PANELIST]
          description: |
            Speaker's role in the session:
            - PRIMARY_SPEAKER: Main presenter
            - CO_SPEAKER: Co-presenter
            - MODERATOR: Panel moderator
            - PANELIST: Panel participant
          example: PRIMARY_SPEAKER
        presentationTitle:
          type: string
          description: Optional speaker-specific presentation title (if different from session title)
          maxLength: 255
          example: Security Auditing Tools Deep Dive
        isConfirmed:
          type: boolean
          description: Whether speaker has confirmed participation
          example: true
    SessionResponse:
      description: |
        Story 1.15a.1b: Session response with embedded speakers
        Same as Session schema but used specifically for API responses
      allOf:
        - $ref: '#/components/schemas/Session'
...
    UpdateEventRequest:
      type: object
      required:
        - title
        - eventNumber
        - date
        - registrationDeadline
        - venueName
        - venueAddress
        - venueCapacity
        - organizerUsername
        - eventType
      properties:
        title:
          type: string
          maxLength: 200
        eventNumber:
          type: integer
        date:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        venueName:
          type: string
          maxLength: 255
        venueAddress:
          type: string
        venueCapacity:
          type: integer
        workflowState:
          $ref: '#/components/schemas/EventWorkflowState'
        organizerUsername:
          type: string
          pattern: '^[a-z]+\.[a-z]+(\.[0-9]+)?$'
          example: john.doe
          description: |
            Username of the event organizer in format "firstname.lastname" or "firstname.lastname.2" for collisions.
            Story 1.16.2: Public API uses meaningful IDs (usernames), not UUIDs.
        currentAttendeeCount:
          type: integer
        publishedAt:
          type: string
          format: date-time
        metadata:
          type: string
        description:
          type: string
          maxLength: 5000
        eventType:
          $ref: '#/components/schemas/EventType'
        themeImageUploadId:
          type: string
          nullable: true
          maxLength: 100
          description: Upload ID from /logos/presigned-url for event theme image
    PatchEventRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        eventNumber:
          type: integer
        date:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        venueName:
          type: string
          maxLength: 255
        venueAddress:
          type: string
        venueCapacity:
          type: integer
        workflowState:
          $ref: '#/components/schemas/EventWorkflowState'
        organizerUsername:
          type: string
          pattern: '^[a-z]+\.[a-z]+(\.[0-9]+)?$'
          example: john.doe
          description: |
            Username of the event organizer in format "firstname.lastname" or "firstname.lastname.2" for collisions.
            Story 1.16.2: Public API uses meaningful IDs (usernames), not UUIDs.
        currentAttendeeCount:
          type: integer
        publishedAt:
          type: string
          format: date-time
        metadata:
          type: string
        description:
          type: string
          maxLength: 5000
        themeImageUploadId:
          type: string
          nullable: true
          maxLength: 100
          description: Upload ID from /logos/presigned-url for event theme image
...
    EventType:
      type: string
      description: |
        Event type identifier (Story 5.1 - Event Type Definition).
        Matches EventType enum from architecture (03-data-architecture.md:547-551).
      enum:
        - FULL_DAY
        - AFTERNOON
        - EVENING
      example: FULL_DAY
...
    ErrorResponse:
      type: object
      description: Standard error response from shared-kernel
      x-java-type: ch.batbern.shared.dto.ErrorResponse
      x-java-type-import: ch.batbern.shared.dto.ErrorResponse
    EventWorkflowState:
      type: string
      description: |
        Event workflow state for 9-step consolidated workflow (reconciliation plan 2025-12-30).
        Defines the current state of an event in the organizer workflow.
        Previous 16-state workflow was consolidated to improve clarity and automation.
      enum:
        - CREATED
        - TOPIC_SELECTION
        - SPEAKER_IDENTIFICATION
        - SLOT_ASSIGNMENT
        - AGENDA_PUBLISHED
        - AGENDA_FINALIZED
        - EVENT_LIVE
        - EVENT_COMPLETED
        - ARCHIVED
      example: SPEAKER_IDENTIFICATION
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              value:
                error: VALIDATION_ERROR
                errorCode: BAD_REQUEST
                message: Invalid filter syntax
                timestamp: 2025-01-15T10:30:00Z
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            eventNotFound:
              value:
                error: RESOURCE_NOT_FOUND
                errorCode: NOT_FOUND
                message: Event not found with id evt-999
                timestamp: 2025-01-15T10:30:00Z
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              value:
                error: INTERNAL_SERVER_ERROR
                errorCode: INTERNAL_ERROR
                message: An unexpected error occurred
                timestamp: 2025-01-15T10:30:00Z
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbiddenError:
              value:
                error: FORBIDDEN
                errorCode: FORBIDDEN
                message: "Insufficient permissions to perform this action"
                timestamp: "2025-01-15T10:30:00Z"
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              value:
                error: UNAUTHORIZED
                errorCode: UNAUTHORIZED
                message: "Authentication token is missing or invalid"
                timestamp: "2025-01-15T10:30:00Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
security:
  - BearerAuth: []